+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/types/sorttypes/AlbumSortType.kt
@@ -8,12 +8,9 @@ fun checkAlbumSortType(sortType: Int, order: Int): String {
     //TODO: **Review this code later**
     val pOrder: String = if (order == 0) " ASC" else " DESC"
     return when (sortType) {
-        0 -> MediaStore.Audio.Albums.DEFAULT_SORT_ORDER + pOrder
-        1 -> MediaStore.Audio.Albums.ALBUM + pOrder
-        2 -> MediaStore.Audio.Albums.ARTIST + pOrder
-        3 -> MediaStore.Audio.Albums.NUMBER_OF_SONGS + pOrder
-        4 -> MediaStore.Audio.Albums.LAST_YEAR
-        5 -> MediaStore.Audio.Albums.FIRST_YEAR
-        else -> MediaStore.Audio.Albums.DEFAULT_SORT_ORDER //
+        0 -> MediaStore.Audio.Albums.ALBUM + pOrder
+        1 -> MediaStore.Audio.Albums.ARTIST + pOrder
+        2 -> MediaStore.Audio.Albums.NUMBER_OF_SONGS + pOrder
+        else -> MediaStore.Audio.Albums.ALBUM //
     }
 }