+++ on_audio_query_platform_interface/lib/details/extensions/models_extensions.dart
@@ -17,18 +17,18 @@ part of on_audio_query_helper;
 /// life easy you can use this method to convert into any model.
 extension OnModelFormatter on List<dynamic> {
   /// Used to convert a [List dynamic] into a [List SongModel].
-  List<SongModel> toSongModel() => this.map((e) => SongModel(e)).toList();
+  List<SongModel> toSongModel() => map((e) => SongModel(e)).toList();
 
   /// Used to convert a [List dynamic] into a [List AlbumModel].
-  List<AlbumModel> toAlbumModel() => this.map((e) => AlbumModel(e)).toList();
+  List<AlbumModel> toAlbumModel() => map((e) => AlbumModel(e)).toList();
 
   /// Used to convert a [List dynamic] into a [List PlaylistModel].
   List<PlaylistModel> toPlaylistModel() =>
-      this.map((e) => PlaylistModel(e)).toList();
+      map((e) => PlaylistModel(e)).toList();
 
   /// Used to convert a [List dynamic] into a [List ArtistModel].
-  List<ArtistModel> toArtistModel() => this.map((e) => ArtistModel(e)).toList();
+  List<ArtistModel> toArtistModel() => map((e) => ArtistModel(e)).toList();
 
   /// Used to convert a [List dynamic] into a [List GenreModel].
-  List<GenreModel> toGenreModel() => this.map((e) => GenreModel(e)).toList();
+  List<GenreModel> toGenreModel() => map((e) => GenreModel(e)).toList();
 }