+++ on_audio_query_platform_interface/lib/details/models/playlist_model.dart
@@ -5,14 +5,10 @@ class PlaylistModel {
   PlaylistModel(this._info);
 
   //The type dynamic is used for both but, the map is always based in [String, dynamic]
-  Map<dynamic, dynamic> _info;
+  final Map<dynamic, dynamic> _info;
 
   ///Return playlist [id]
   int get id => _info["_id"];
-
-  /// Deprecated after [2.0.0].
-  @Deprecated("Use [album] instead")
-  String? get playlistName => playlist;
 
   ///Return playlist [playlist]
   String get playlist => _info["name"];
@@ -25,6 +21,9 @@ class PlaylistModel {
 
   ///Return playlist [dateModified]
   int? get dateModified => _info["date_modified"];
+
+  ///Return playlist [numOfSongs]
+  int get numOfSongs => _info["num_of_songs"];
 
   /// Return a map with all [keys] and [values] from specific playlist.
   Map get getMap => _info;