+++ on_audio_query_platform_interface/lib/details/on_audio_query_helper.dart
@@ -1,7 +1,5 @@
 library on_audio_query_helper;
 
-import 'dart:typed_data';
-
 //Extensions
 part 'extensions/models_extensions.dart';
 
@@ -24,7 +22,5 @@ part 'types/sort_types/genre_sort_type.dart';
 part 'types/order_type.dart';
 part 'types/artwork_type.dart';
 part 'types/audios_from_type.dart';
-part 'types/audios_only_type.dart';
 part 'types/with_filters_type.dart';
 part 'types/uri_type.dart';
-part 'types/songs_by_type.dart';