+++ on_audio_query_platform_interface/lib/details/types/sort_types/playlist_sort_type.dart
@@ -1,17 +1,20 @@
+// ignore_for_file: constant_identifier_names
+
 part of on_audio_query_helper;
 
 /// Defines the sort type used for [queryPlaylists].
 enum PlaylistSortType {
-  /// [DEFAULT] will return playlist using [name] as sort param.
-  DEFAULT,
-
-  /// [DATA_ADDED] will return playlist based in [data_added].
-  DATA_ADDED,
-
   /// [PLAYLIST] will return playlist based in [playlist].
   PLAYLIST,
 
-  /// Deprecated after [2.0.0].
+  /// [DATE_ADDED] will return playlist based in [data_added].
+  DATE_ADDED,
+
+  /// Deprecated after [2.3.0].
   @Deprecated("Use [PLAYLIST] instead")
-  PLAYLIST_NAME,
+  DEFAULT,
+
+  /// Deprecated after [2.3.0].
+  @Deprecated("Use [DATE_ADDED] instead")
+  DATA_ADDED,
 }