+++ on_audio_query_platform_interface/lib/details/types/sort_types/song_sort_type.dart
@@ -1,9 +1,11 @@
+// ignore_for_file: constant_identifier_names
+
 part of on_audio_query_helper;
 
 /// Defines the sort type used for [querySongs] and [queryAudios].
 enum SongSortType {
-  ///[DEFAULT] will return song list using [title] as sort param.
-  DEFAULT,
+  ///[TITLE] will return song list using [title] as sort param.
+  TITLE,
 
   ///[ARTIST] will return song list based in artist [name].
   ARTIST,
@@ -11,19 +13,23 @@ enum SongSortType {
   ///[ALBUM] will return song list based in album [name].
   ALBUM,
 
-  /// Deprecated after [2.0.0].
-  @Deprecated("This method will be removed soon")
-  YEAR,
-
   ///[DURATION] will return song list based in song [duration].
   DURATION,
 
-  ///[DATA_ADDED] will return song list based in [data_added].
-  DATA_ADDED,
+  ///[DATE_ADDED] will return song list based in [data_added].
+  DATE_ADDED,
 
   ///[SIZE] will return song list based in song [size].
   SIZE,
 
   ///[DISPLAY_NAME] will return song list based in song [display_name].
   DISPLAY_NAME,
+
+  /// Deprecated after [2.3.0].
+  @Deprecated("Use [TITLE] instead")
+  DEFAULT,
+
+  /// Deprecated after [2.3.0].
+  @Deprecated("Use [DATE_ADDED] instead")
+  DATA_ADDED,
 }