+++ on_audio_query_platform_interface/lib/details/types/with_filters_type.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: constant_identifier_names
+
 part of on_audio_query_helper;
 
 /// Defines the type of Filter.
@@ -65,38 +67,22 @@ enum AlbumsArgs {
 
   /// Uses album [ARTIST] as filter.
   ARTIST,
-
-  /// Deprecated after [2.0.0].
-  @Deprecated("Use [ALBUM] instead")
-  ALBUM_NAME,
 }
 
 /// Args types for Playlists.
 enum PlaylistsArgs {
   /// Uses [PLAYLIST] as filter.
   PLAYLIST,
-
-  /// Deprecated after [2.0.0].
-  @Deprecated("Use [PLAYLIST] instead")
-  PLAYLIST_NAME,
 }
 
 /// Args types for Artists.
 enum ArtistsArgs {
   /// Uses [ARTIST] as filter.
   ARTIST,
-
-  /// Deprecated after [2.0.0].
-  @Deprecated("Use [ARTIST] instead")
-  ARTIST_NAME,
 }
 
 /// Args types for Genres.
 enum GenresArgs {
   /// Uses [GENRE] as filter.
   GENRE,
-
-  /// Deprecated after [2.0.0].
-  @Deprecated("Use [GENRE] instead")
-  GENRE_NAME,
 }